Skipped File::symlink test on MSWin32 Made Time::time2str default to UTC For MSWin32 changed Functions::abs_path to return backslashes

+++ t/20file.t
@@ -35,13 +35,15 @@ use Exception::Class ( q(TestException) => { fields => [ qw(arg1 arg2) ] } );
 use Class::Usul;
 use Class::Usul::File;
 
-my $cu = Class::Usul->new( config       => {
-                              appclass  => q(Class::Usul),
-                              home      => catdir( qw(lib Class Usul) ),
-                              localedir => catdir( qw(t locale) ),
-                              tempdir   => q(t), },
-                           debug        => 0,
-                           log          => Logger->new, );
+my $osname = lc $OSNAME;
+my $cu     = Class::Usul->new
+   ( config       => {
+      appclass  => q(Class::Usul),
+      home      => catdir( qw(lib Class Usul) ),
+      localedir => catdir( qw(t locale) ),
+      tempdir   => q(t), },
+     debug        => 0,
+     log          => Logger->new, );
 
 my $cuf = Class::Usul::File->new( builder => $cu );
 
@@ -69,11 +71,13 @@ unlink catfile( qw(t ipc_srlock.shm) );
 
 is $cuf->status_for( $tf )->{size}, 237, 'status_for';
 
-my $symlink = catfile( qw(t symlink) );
+if ($osname ne 'mswin32') {
+   my $symlink = catfile( qw(t symlink) );
 
-$cuf->symlink( q(t), q(test.xml), [ qw(t symlink) ] );
+   $cuf->symlink( q(t), q(test.xml), [ qw(t symlink) ] );
 
-ok -e $symlink, 'symlink'; -e _ and unlink $symlink;
+   ok -e $symlink, 'symlink'; -e _ and unlink $symlink;
+}
 
 my $tempfile = $cuf->tempfile;
